accounts: don't remove accounts while device is in safe mode
Userspace apps are not available in safe mode, so authenticators are not available when
the system starts and checks all accounts. Just ignore these accounts and not add them
to the list of accounts.

JIRA: CYAN-4074
Issue: https://jira.cyanogenmod.org/browse/CYAN-4074

+++ services/java/com/android/server/accounts/AccountManagerService.java
@@ -111,6 +111,7 @@ public class AccountManagerService
     private static final int DATABASE_VERSION = 5;
 
     private final Context mContext;
+    private boolean mSafeMode;
 
     private final PackageManager mPackageManager;
     private UserManager mUserManager;
@@ -283,6 +284,10 @@ public class AccountManagerService
         }
     }
 
+    public void setSafeMode(boolean safeMode) {
+        mSafeMode = safeMode;
+    }
+
     public void systemReady() {
     }
 
@@ -355,6 +360,7 @@ public class AccountManagerService
      */
     private void validateAccountsInternal(
             UserAccounts accounts, boolean invalidateAuthenticatorCache) {
+System.out.println("JRC::validateAccountsInternal: safeMode: " + mSafeMode);
         if (invalidateAuthenticatorCache) {
             mAuthenticatorCache.invalidateCache(accounts.userId);
         }
@@ -381,10 +387,15 @@ public class AccountManagerService
                     final String accountName = cursor.getString(2);
 
                     if (!knownAuth.contains(AuthenticatorDescription.newKey(accountType))) {
-                        Slog.w(TAG, "deleting account " + accountName + " because type "
-                                + accountType + " no longer has a registered authenticator");
-                        db.delete(TABLE_ACCOUNTS, ACCOUNTS_ID + "=" + accountId, null);
-                        accountDeleted = true;
+                        if (mSafeMode) {
+                            Slog.w(TAG, "ignoring account " + accountName + " because type "
+                                    + accountType + " has not registered authenticator (safemode)");
+                        } else {
+                            Slog.w(TAG, "deleting account " + accountName + " because type "
+                                    + accountType + " no longer has a registered authenticator");
+                            db.delete(TABLE_ACCOUNTS, ACCOUNTS_ID + "=" + accountId, null);
+                            accountDeleted = true;
+                        }
                         final Account account = new Account(accountName, accountType);
                         accounts.userDataCache.remove(account);
                         accounts.authTokenCache.remove(account);
